Moved TypeHelper to sole usage
